corrigido KNN().createENodes()

+++ src/knearestneighbor/Node.java
@@ -3,16 +3,16 @@ package knearestneighbor;
 public class Node {
     
     private Type type;
-    private double area;
+    private int area;
     private int rooms;
     
-    public Node(Type type, double area, int rooms) {
+    public Node(Type type, int area, int rooms) {
         this.type = type;
         this.area = area;
         this.rooms = rooms;
     }
     
-    public Node(double area, int rooms) {
+    public Node(int area, int rooms) {
         this.area = area;
         this.rooms = rooms;
     }
@@ -26,11 +26,11 @@ public class Node {
         this.type = type;
     }
 
-    public double getArea() {
+    public int getArea() {
         return area;
     }
 
-    public void setArea(double area) {
+    public void setArea(int area) {
         this.area = area;
     }
